+++ setup.cfg
@@ -123,8 +123,9 @@ test_petab =
 
 [options.entry_points]
 console_scripts =
-    abc-server = pyabc.visserver.server:run_app
+    abc-server-flask = pyabc.visserver.server_flask:run_app
     abc-server-dash = pyabc.visserver.server_dash:run_app
+    abc-server = pyabc.visserver.server_dash:run_app
     abc-redis-worker = pyabc.sampler.redis_eps.cli:work
     abc-redis-manager = pyabc.sampler.redis_eps.cli:manage
     abc-export = pyabc.storage.db_export:main
